+++ data/LictData.txt
@@ -1,1 +1,3 @@
-DEADLINE | 1 | tjoef | 2019-10-14 1400
+DEADLINE | 0 | tjoef | 2019-10-14 1400
+DEADLINE | 0 | aksmd | 2019-10-01
+TODO | 0 | aslmdf
