+++ src/main/java/lict/task/Task.java
@@ -1,5 +1,9 @@
+package lict.task;
+
+import lict.LictException;
+
 /**
- * The abstract {@code Task} class represents a task with a description and a status indicating whether it is done.
+ * The abstract {@code lict.task.Task} class represents a task with a description and a status indicating whether it is done.
  * It provides methods to manage the task's completion status and to convert it to a string representation or data representation for saving and loading.
  */
