+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 __________________________________
-Hello! I'm Lict
+Hello! I'm lict.lict
 What can I do for you?
 __________________________________
 __________________________________
